Cleanup multiple -> ledger
