+++ src/host/test/ledger.cpp
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "../ds/serialized.h"
-#include "../multiple_ledger.h"
+#include "../ledger.h"
 
 #include <doctest/doctest.h>
 #include <string>
@@ -89,13 +89,13 @@ void verify_framed_entries_range(
   REQUIRE(idx == to + 1);
 }
 
-void read_entry_from_ledger(asynchost::MultipleLedger& ledger, size_t idx)
+void read_entry_from_ledger(asynchost::Ledger& ledger, size_t idx)
 {
   REQUIRE(TestLedgerEntry(ledger.read_entry(idx).value()).value() == idx);
 }
 
 void read_entries_range_from_ledger(
-  asynchost::MultipleLedger& ledger, size_t from, size_t to)
+  asynchost::Ledger& ledger, size_t from, size_t to)
 {
   verify_framed_entries_range(
     ledger.read_framed_entries(from, to).value(), from, to);
@@ -107,12 +107,11 @@ void read_entries_range_from_ledger(
 class TestEntrySubmitter
 {
 private:
-  asynchost::MultipleLedger& ledger;
+  asynchost::Ledger& ledger;
   size_t last_idx;
 
 public:
-  TestEntrySubmitter(
-    asynchost::MultipleLedger& ledger, size_t initial_last_idx = 0) :
+  TestEntrySubmitter(asynchost::Ledger& ledger, size_t initial_last_idx = 0) :
     ledger(ledger),
     last_idx(initial_last_idx)
   {}
@@ -185,12 +184,12 @@ TEST_CASE("Regular chunking")
   INFO("Cannot create a ledger with a chunk threshold of 0");
   {
     size_t chunk_threshold = 0;
-    REQUIRE_THROWS(asynchost::MultipleLedger(ledger_dir, wf, chunk_threshold));
+    REQUIRE_THROWS(asynchost::Ledger(ledger_dir, wf, chunk_threshold));
   }
 
   size_t chunk_threshold = 30;
   size_t entries_per_chunk = get_entries_per_chunk(chunk_threshold);
-  asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+  asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
   TestEntrySubmitter entry_submitter(ledger);
 
   size_t end_of_first_chunk_idx = 0;
@@ -305,7 +304,7 @@ TEST_CASE("Truncation")
   fs::remove_all(ledger_dir);
 
   size_t chunk_threshold = 30;
-  asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+  asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
   TestEntrySubmitter entry_submitter(ledger);
 
   size_t chunk_count = 3;
@@ -380,7 +379,7 @@ TEST_CASE("Commit")
   fs::remove_all(ledger_dir);
 
   size_t chunk_threshold = 30;
-  asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+  asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
   TestEntrySubmitter entry_submitter(ledger);
 
   size_t chunk_count = 3;
@@ -468,7 +467,7 @@ TEST_CASE("Restore existing ledger")
   {
     INFO("Initialise first ledger with complete chunks");
     {
-      asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+      asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
       TestEntrySubmitter entry_submitter(ledger);
 
       end_of_first_chunk_idx =
@@ -477,7 +476,7 @@ TEST_CASE("Restore existing ledger")
       last_idx = chunk_count * end_of_first_chunk_idx;
     }
 
-    asynchost::MultipleLedger ledger2(ledger_dir, wf, chunk_threshold);
+    asynchost::Ledger ledger2(ledger_dir, wf, chunk_threshold);
     read_entries_range_from_ledger(ledger2, 1, last_idx);
 
     // Restored ledger can be written to
@@ -498,7 +497,7 @@ TEST_CASE("Restore existing ledger")
   {
     INFO("Initialise first ledger with truncation");
     {
-      asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+      asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
       TestEntrySubmitter entry_submitter(ledger);
 
       end_of_first_chunk_idx =
@@ -509,7 +508,7 @@ TEST_CASE("Restore existing ledger")
       number_of_ledger_files = number_of_files_in_ledger_dir();
     }
 
-    asynchost::MultipleLedger ledger2(ledger_dir, wf, chunk_threshold);
+    asynchost::Ledger ledger2(ledger_dir, wf, chunk_threshold);
     read_entries_range_from_ledger(ledger2, 1, last_idx);
 
     TestEntrySubmitter entry_submitter(ledger2, last_idx);
@@ -525,7 +524,7 @@ TEST_CASE("Restore existing ledger")
     size_t committed_idx = 0;
     INFO("Initialise first ledger with committed chunks");
     {
-      asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+      asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
       TestEntrySubmitter entry_submitter(ledger);
 
       end_of_first_chunk_idx =
@@ -537,7 +536,7 @@ TEST_CASE("Restore existing ledger")
       ledger.commit(committed_idx);
     }
 
-    asynchost::MultipleLedger ledger2(ledger_dir, wf, chunk_threshold);
+    asynchost::Ledger ledger2(ledger_dir, wf, chunk_threshold);
     read_entries_range_from_ledger(ledger2, 1, last_idx);
 
     // Restored ledger cannot be truncated before last idx of last committed
@@ -553,7 +552,7 @@ TEST_CASE("Restore existing ledger")
   {
     INFO("Initialise first ledger with committed chunks");
     {
-      asynchost::MultipleLedger ledger(ledger_dir, wf, chunk_threshold);
+      asynchost::Ledger ledger(ledger_dir, wf, chunk_threshold);
       TestEntrySubmitter entry_submitter(ledger);
 
       end_of_first_chunk_idx =
@@ -565,7 +564,7 @@ TEST_CASE("Restore existing ledger")
 
     INFO("Restore new ledger with twice the chunking threshold");
     {
-      asynchost::MultipleLedger ledger2(ledger_dir, wf, 2 * chunk_threshold);
+      asynchost::Ledger ledger2(ledger_dir, wf, 2 * chunk_threshold);
       read_entries_range_from_ledger(ledger2, 1, last_idx);
 
       TestEntrySubmitter entry_submitter(ledger2, last_idx);
@@ -580,7 +579,7 @@ TEST_CASE("Restore existing ledger")
 
     INFO("Restore new ledger with half the chunking threshold");
     {
-      asynchost::MultipleLedger ledger2(ledger_dir, wf, chunk_threshold / 2);
+      asynchost::Ledger ledger2(ledger_dir, wf, chunk_threshold / 2);
       read_entries_range_from_ledger(ledger2, 1, last_idx);
 
       TestEntrySubmitter entry_submitter(ledger2, last_idx);
@@ -611,7 +610,7 @@ TEST_CASE("Limit number of open files")
   size_t chunk_threshold = 30;
   size_t chunk_count = 5;
   size_t max_read_cache_size = 2;
-  asynchost::MultipleLedger ledger(
+  asynchost::Ledger ledger(
     ledger_dir, wf, chunk_threshold, max_read_cache_size);
   TestEntrySubmitter entry_submitter(ledger);
 
@@ -679,7 +678,7 @@ TEST_CASE("Limit number of open files")
   INFO("Still possible to recover a new ledger");
   {
     initial_number_fd = number_open_fd();
-    asynchost::MultipleLedger ledger2(
+    asynchost::Ledger ledger2(
       ledger_dir, wf, chunk_threshold, max_read_cache_size);
 
     // Committed files are not open for write
